Stores registrant with InvitationKey, and uses it to prevent key reuse.
This is a fix for issue 3:
http://bitbucket.org/david/django-invitation/issue/3/expire-used-invitations

A key is usable if it is not expired (as before), and if also no one
has registered using the key.

This change adds the column 'registrant' to InvitationKey, and so
requires changing your database schema.  The column can be NULL, so
it's a pretty simple addition.

(I chose to use 'registrant' rather than 'to_user' (which would parallel
'from_user') because technically we don't know who the invitation key was
sent to.  All we know is that this particular user used the key to register.)

The implementation uses profile_callback, which more tightly couples
django-invitation to version 0.7 of django-registration.  (It already
had this requirement based on the function signatures, but now we're
actually using the extra profile_callback argument.)